+++ admin/26_deploy-mysql-provider.ps1
@@ -3,6 +3,11 @@
 [securestring]$PfxPass = $Global:VMPassword,
 $MySQLRPadmin = $Global:MySQLRPAdmin
 )
+if (!$Global:SubscriptionID)
+    {
+    Write-Warning -Message "You Have not Configured a SubscriptionID, did you run 99_bootstrap.ps1 ?"
+    break
+}
 push-location
 $MYSQL_DIR = "C:\Temp\MySQL"
 Remove-Item $MYSQL_DIR -Force -Recurse -ErrorAction SilentlyContinue -Confirm:$false | out-null
